# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Business Hub API
  description: |
    All-in-one business management platform API.
    Covers leads, jobs, content, task boards, templates, skill mastery,
    AI calling, agents, and campaigns.
  version: 1.0.0

servers:
  - url: http://localhost:3002
    description: Development

security:
  - userIdHeader: []

paths:
  # ── Auth ──
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
                name: { type: string }
      responses:
        '201': { description: User created }
        '400': { description: Validation error or email exists }

  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  email: { type: string }
                  name: { type: string }
        '401': { description: Invalid credentials }

  /api/auth/profile:
    get:
      tags: [Auth]
      summary: Get current user profile
      responses:
        '200': { description: User profile }

  # ── Leads ──
  /api/leads:
    get:
      tags: [Leads]
      summary: List all leads
      responses:
        '200': { description: Array of leads }
    post:
      tags: [Leads]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email]
              properties:
                name: { type: string }
                email: { type: string }
                phone: { type: string }
                company: { type: string }
                status: { type: string, enum: [new, contacted, qualified, lost, won] }
      responses:
        '201': { description: Lead created }

  /api/leads/{id}:
    get:
      tags: [Leads]
      summary: Get lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Lead details }
    patch:
      tags: [Leads]
      summary: Update lead
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Updated lead }
    delete:
      tags: [Leads]
      summary: Delete lead
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Lead deleted }

  # ── Jobs ──
  /api/jobs:
    get:
      tags: [Jobs]
      summary: List all jobs
      responses:
        '200': { description: Array of jobs }
    post:
      tags: [Jobs]
      summary: Create a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, company]
              properties:
                title: { type: string }
                company: { type: string }
                status: { type: string }
                url: { type: string }
                salary: { type: string }
                notes: { type: string }
      responses:
        '201': { description: Job created }

  /api/jobs/{id}:
    get:
      tags: [Jobs]
      summary: Get job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Job details }
    patch:
      tags: [Jobs]
      summary: Update job
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Updated job }
    delete:
      tags: [Jobs]
      summary: Delete job
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Job deleted }

  # ── Calls ──
  /api/calls:
    get:
      tags: [Calls]
      summary: List calls with filters
      parameters:
        - name: leadId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer }
        - name: offset
          in: query
          schema: { type: integer }
      responses:
        '200': { description: Paginated call list }

  /api/calls/stats:
    get:
      tags: [Calls]
      summary: Call analytics / stats
      responses:
        '200':
          description: Call statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalCalls: { type: integer }
                  todayCalls: { type: integer }
                  bookedCount: { type: integer }
                  avgDuration: { type: number }
                  conversionRate: { type: number }
                  outcomeBreakdown: { type: object }
                  dailyVolume:
                    type: array
                    items:
                      type: object
                      properties:
                        date: { type: string }
                        count: { type: integer }
                        booked: { type: integer }

  /api/calls/initiate:
    post:
      tags: [Calls]
      summary: Start an outbound call
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                leadId: { type: string }
                scriptId: { type: string }
      responses:
        '201': { description: Call initiated }

  # ── Agents ──
  /api/agents:
    get:
      tags: [Agents]
      summary: List AI agents
      responses:
        '200': { description: Array of agents }

  /api/agents/{id}:
    get:
      tags: [Agents]
      summary: Get agent details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Agent details with flow config }

  # ── Campaigns ──
  /api/campaigns:
    get:
      tags: [Campaigns]
      summary: List campaigns
      responses:
        '200': { description: Array of campaigns }
    post:
      tags: [Campaigns]
      summary: Create campaign
      responses:
        '201': { description: Campaign created }

  /api/campaigns/analytics:
    get:
      tags: [Campaigns]
      summary: Campaign analytics
      responses:
        '200': { description: Analytics data }

  # ── Health ──
  /api/health:
    get:
      tags: [System]
      summary: Health check
      security: []
      responses:
        '200':
          description: System health
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
                  uptime: { type: string, example: "3600s" }
                  database: { type: string, enum: [connected, disconnected] }
                  memory:
                    type: object
                    properties:
                      rss: { type: string }
                      heap: { type: string }
                  wsClients: { type: integer }

components:
  securitySchemes:
    userIdHeader:
      type: apiKey
      in: header
      name: x-user-id
      description: User ID from login response

tags:
  - name: Auth
    description: Authentication (no auth required)
  - name: Leads
    description: Lead management
  - name: Jobs
    description: Job tracking
  - name: Calls
    description: AI calling system
  - name: Agents
    description: AI agent management
  - name: Campaigns
    description: Outreach campaigns
  - name: System
    description: System health and diagnostics